requirements.txt
altgraph==0.17.4
appscript==1.2.5
av==9.1.1
contourpy==1.2.0
cycler==0.12.1
docx2pdf==0.1.8
fonttools==4.49.0
kiwisolver==1.4.5
lxml==5.1.1
macholib==1.16.3
matplotlib==3.8.3
mido==1.3.0
numpy==1.26.4
packaging==23.2
Pillow==10.1.0
PyAudio==0.2.14
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.11
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-rtmidi==1.5.8
six==1.16.0
tkPDFViewer2==0.1
tkvideoplayer==2.5
tqdm==4.66.2
ttkbootstrap==1.10.1
typing_extensions==4.10.0
